+++ src/sc-preferences-page.h
@@ -13,6 +13,6 @@ struct _ScPreferencesPageClass
   AdwPreferencesPageClass parent_class;
 };
 
-uint16_t sc_preferences_page_get_control_id_offset (ScPreferencesPage *self);
+uint32_t sc_preferences_page_get_control_id_offset (ScPreferencesPage *self);
 
 G_END_DECLS
